endpointmanager: signal when work is done
Provide new ways of signaling when various operations are completed by the
endpointmanager for the following cases:

* Removal of endpoint: when an endpoint is removed, all references to it must
also be removed from the policy repository in the Daemon as well. Block on
releasing the endpoint ID until it is removed from all caches in the policy
repository, but do so in an asynchronous manner so that `endpointmanager.Remove`
does not get stuck for a long period of time waiting for said work to complete.
* Provide new functions to indicate when regenerations for a set of endpoints
have been enqueued for each endpoint. This is useful for ensuring that a given
regeneration is queued up before some other event, e.g., incrementing the policy
revision for the endpoint for a later revision than the regeneration which is
being queued.

Signed-off by: Ian Vernon <ian@cilium.io>